Pull `prime_fmod_size` into its own dedicated header, update #include list for fca.hpp and prime_fmod_test.hpp
